Move api into api/ow4 to prepare for Surkle
