+++ src/actions/store.ts
@@ -1,7 +1,8 @@
 import * as authA from "actions/auth";
 import * as userA from "actions/user";
+import ow4 from "api/ow4";
 import { basketPrice } from "helpers/store";
-import { IBasketItem, IItem } from "models/item";
+import { IItem } from "models/item";
 import { RootState } from "reducers/root-reducer";
 import { ThunkDispatch } from "redux-thunk";
 import { ActionType, createAction } from "typesafe-actions";
@@ -24,25 +25,6 @@ export const purchaseRequest = createAction("store/PURCHASE_REQUEST");
 export const purchaseSuccess = createAction("store/PURCHASE_SUCCESS");
 export const purchaseFailure = createAction("store/PURCHASE_FAILURE");
 
-interface IAPIBasketItem {
-    object_id: string;
-    amount: number;
-}
-
-function prepareBasket(basket: IBasketItem[]): IAPIBasketItem[] {
-    const count: {[id: string]: number} = {};
-
-    basket.forEach((bItem) => {
-        if (!count[bItem.id]) {
-            count[bItem.id] = 1;
-        } else {
-            count[bItem.id]++;
-        }
-    });
-
-    return Object.keys(count).map((k) => ({ object_id: k, amount: count[k] }));
-}
-
 export function purchase() {
     const callback = async (
         dispatch: ThunkDispatch<RootState, void, StoreAction | userA.UserAction | authA.AuthAction>,
@@ -55,17 +37,7 @@ export function purchase() {
             dispatch(purchaseSuccess());
             dispatch(authA.logout());
         } else {
-            const res = await fetch(`https://online.ntnu.no/api/v1/orderline/`, {
-                body: JSON.stringify({
-                    orders: prepareBasket(getStore().store.basket),
-                    user: getStore().auth.id,
-                }),
-                headers: {
-                    "Authorization": `Bearer ${getStore().auth.token}`,
-                    "Content-Type": "application/json",
-                },
-                method: "POST",
-            });
+            const res = await ow4.purchase(getStore());
 
             if (res.status === 401) {
                 authA.authenticate()(dispatch);
